rename: 파일 위치 변경
filterNav의 위치를 변경
